feat: 보드 좋아요 연동

+++ src/pages/news/ScrapButton.jsx
@@ -1,5 +1,7 @@
 import apiClient from "../../api/axios";
 import { FaBookmark, FaRegBookmark } from "react-icons/fa";
+import { toast } from "sonner";
+import { useStore } from "../../store/useUserStore";
 
 export default function NewsLikeButton({
   newsId,
@@ -8,14 +10,15 @@ export default function NewsLikeButton({
   scrapCount,
   setScrapCount,
 }) {
-
-
-
-
+  const logId = useStore((state) => state.userId);
   const handleClick = async () => {
+    if (!logId) {
+      toast.warning("로그인 후 이용 가능합니다.");
+      return;
+    }
     if (isScrap) {
       try {
-        const response = await apiClient.delete(`/api/news/${newsId}/scrap`);
+        const res = await apiClient.delete(`/api/news/${newsId}/scrap`);
         setIsScrap(false);
         setScrapCount(scrapCount - 1);
       } catch (error) {
@@ -23,7 +26,7 @@ export default function NewsLikeButton({
       }
     } else {
       try {
-        const response = await apiClient.post(`/api/news/${newsId}/scrap`);
+        const res = await apiClient.post(`/api/news/${newsId}/scrap`);
         setIsScrap(true);
         setScrapCount(scrapCount + 1);
       } catch (error) {
@@ -33,7 +36,10 @@ export default function NewsLikeButton({
   };
 
   return (
-    <button onClick={handleClick} className="px-3 py-2 rounded-lg flex items-center space-x-2 transition-all duration-300 cursor-pointer">
+    <button
+      onClick={handleClick}
+      className="px-3 py-2 rounded-lg flex items-center space-x-2 transition-all duration-300 cursor-pointer"
+    >
       {isScrap ? (
         <FaBookmark className="w-5 h-5 text-orange-500 transform transition-transform duration-300 hover:scale-110" /> // 찬 스크랩
       ) : (
@@ -41,11 +47,10 @@ export default function NewsLikeButton({
       )}
 
       <span
-        className={`text-sm font-medium ${isScrap ? "text-orange-500" : "text-gray-500"
-          } transition-colors duration-300`}
-      >
-
-      </span>
+        className={`text-sm font-medium ${
+          isScrap ? "text-orange-500" : "text-gray-500"
+        } transition-colors duration-300`}
+      ></span>
     </button>
   );
 }
